committed If you spend too much time thinking about a thing, you'll never get it done

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,6 +1,4 @@
 #include "function_pointers.h"
-#include <stdio.h>
-#include <stddef.h>
 
 /**
  * array_iterator - a function that executes a function given as
@@ -12,7 +10,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int a;
+	unsigned int a;
 
 	if (array == NULL)
 		return;
